+++ packages/network-controller/src/NetworkController.ts
@@ -456,8 +456,6 @@ export type NetworkControllerNetworkRemovedEvent = {
  * @param payload.error - The last error produced by the last failover in the
  * endpoint chain.
  * @param payload.networkClientId - The target network's client ID.
- * @param payload.rpcMethodName - The JSON-RPC method that was being executed
- * when the chain became unavailable.
  */
 export type NetworkControllerRpcEndpointChainUnavailableEvent = {
   type: 'NetworkController:rpcEndpointChainUnavailable';
@@ -466,7 +464,6 @@ export type NetworkControllerRpcEndpointChainUnavailableEvent = {
       chainId: Hex;
       error: unknown;
       networkClientId: NetworkClientId;
-      rpcMethodName: string;
     },
   ];
 };
@@ -487,8 +484,6 @@ export type NetworkControllerRpcEndpointChainUnavailableEvent = {
  * @param payload.error - The last error produced by the endpoint.
  * @param payload.networkClientId - The target network's client ID.
  * @param payload.primaryEndpointUrl - The endpoint chain's primary URL.
- * @param payload.rpcMethodName - The JSON-RPC method that was being executed
- * when the endpoint became unavailable.
  */
 export type NetworkControllerRpcEndpointUnavailableEvent = {
   type: 'NetworkController:rpcEndpointUnavailable';
@@ -499,7 +494,6 @@ export type NetworkControllerRpcEndpointUnavailableEvent = {
       error: unknown;
       networkClientId: NetworkClientId;
       primaryEndpointUrl: string;
-      rpcMethodName: string;
     },
   ];
 };